refactor(web-app): asset pages cleanup common functionalities
